[WIP] TreeBrowserBundle refactoring
This PR refactors the TreeBrowserBundle to make it more generic. All parameters needed to control the javascript tree are taken from the tree. As a side effect, this will (hopefully) properly implement, moving and reordering of nodes, allow for custom icons and valid_children.

Once merged there can be many different trees in the application and we can easily build a pure phpcr browser (or you could build a server side filesystem browser, which is my refactoring test case ...).

This is work in progress, especially the PhpcrOdmTree is not adapted to that yet.